hh h l ll for unsigned int

+++ LIBFT/ft_printf/ft_printf.h
@@ -75,7 +75,7 @@ int convert_uint(va_list args, int fd, t_flags *flags);
 
 int convert_oct(va_list args, int fd, t_flags *flags);
 
-int pad_uint(unsigned int number, t_flags *flags, int fd);
+int pad_uint(uintmax_t number, t_flags *flags, int fd);
 
 int ft_unbrlen(uintmax_t n);
 #endif
